refactor : add correct importations and move status code to domain layer
