+++ src/application/activity/CreateActivityService.ts
@@ -1,6 +1,6 @@
 import { ILogger } from "../../domain/Logger";
 import { AppError } from "../errors/errors";
-import { HttpStatusCode } from "../../infraestructure/utils/HttpStatusCode";
+import { HttpStatusCode } from "../../domain/HttpStatusCode";
 import { Activity } from "../../domain/Activity";
 import { IActivityRepository } from "../../domain/activity/ActivityRepository";
 
@@ -11,7 +11,7 @@ export interface ICreateActivityService {
 export class CreateActivityService implements ICreateActivityService {
   constructor(
     readonly activityRepository: IActivityRepository,
-    readonly logger: ILogger,
+    readonly logger: ILogger
   ) {}
   async invoke(data: Activity): Promise<Activity> {
     try {
@@ -20,11 +20,11 @@ export class CreateActivityService implements ICreateActivityService {
       return activity;
     } catch (error) {
       this.logger.error(
-        `Some error has been ocurred trying create a new activity ${error}`,
+        `Some error has been ocurred trying create a new activity ${error}`
       );
       throw new AppError(
         "Internal server error",
-        HttpStatusCode.InternalServerError,
+        HttpStatusCode.InternalServerError
       );
     }
   }
