+++ src/application/activity/GetActivityService.ts
@@ -1,6 +1,6 @@
 import { Activity } from "../../domain/Activity";
 import { AppError } from "../errors/errors";
-import { HttpStatusCode } from "../../infraestructure/utils/HttpStatusCode";
+import { HttpStatusCode } from "../../domain/HttpStatusCode";
 import { IActivityRepository } from "../../domain/activity/ActivityRepository";
 import { ILogger } from "../../domain/Logger";
 
@@ -15,18 +15,18 @@ export interface IGetActivity {
 export class GetActivityService implements IGetActivity {
   constructor(
     readonly activityRepository: IActivityRepository,
-    readonly logger: ILogger,
+    readonly logger: ILogger
   ) {}
   async invoke({ id }: GetActivityParam): Promise<Activity> {
     try {
       return await this.activityRepository.findById(id);
     } catch (error) {
       this.logger.error(
-        `Some error has been ocurred trying retrieve an activity ${error}`,
+        `Some error has been ocurred trying retrieve an activity ${error}`
       );
       throw new AppError(
         "Internal server error",
-        HttpStatusCode.InternalServerError,
+        HttpStatusCode.InternalServerError
       );
     }
   }
