+++ src/application/users/CreateUserService.ts
@@ -3,14 +3,14 @@ import { HashService } from "../../domain/HashService";
 import { ILogger } from "../../domain/Logger";
 import { IUsersRepository } from "../../domain/users/UsersRepository";
 import { AppError, UserAlreadyExist } from "../errors/errors";
-import { HttpStatusCode } from "../../infraestructure/utils/HttpStatusCode";
+import { HttpStatusCode } from "../../domain/HttpStatusCode";
 import { Users } from "../../domain/Users";
 
 export class CreateUserService implements ICreateUserService {
   constructor(
     readonly usersRepository: IUsersRepository,
     readonly bcrypt: HashService,
-    readonly logger: ILogger,
+    readonly logger: ILogger
   ) {}
 
   async invoke(user: Users): Promise<Users> {
@@ -30,11 +30,11 @@ export class CreateUserService implements ICreateUserService {
       });
     } catch (error) {
       this.logger.error(
-        `Some Internal server error has been ocurred trying create a new user : ${error}`,
+        `Some Internal server error has been ocurred trying create a new user : ${error}`
       );
       throw new AppError(
         "Internal server error",
-        HttpStatusCode.InternalServerError,
+        HttpStatusCode.InternalServerError
       );
     }
   }
