+++ src/presentation/controllers/activity/DeleteActivityController.ts
@@ -1,5 +1,5 @@
 import { Request } from "express";
-import { HttpStatusCode } from "../../../infraestructure/utils/HttpStatusCode";
+import { HttpStatusCode } from "../../../domain/HttpStatusCode";
 import { z } from "zod";
 import { Controller, HttpResponse } from "../../../domain/Controller";
 import { IDeleteActivity } from "../../../application/activity/DeleteActivity";
