+++ src/presentation/controllers/activity/UpdateActivityController.ts
@@ -1,5 +1,5 @@
 import { Request } from "express";
-import { HttpStatusCode } from "../../../infraestructure/utils/HttpStatusCode";
+import { HttpStatusCode } from "../../../domain/HttpStatusCode";
 import { z } from "zod";
 import { IUpdateActivity } from "../../../application/activity/UpdateActivity";
 import { Controller, HttpResponse } from "../../../domain/Controller";
