+++ src/presentation/controllers/authentication/LoginController.ts
@@ -1,5 +1,5 @@
 import { Request } from "express";
-import { HttpStatusCode } from "../../../infraestructure/utils/HttpStatusCode";
+import { HttpStatusCode } from "../../../domain/HttpStatusCode";
 import { z } from "zod";
 import { LoginService } from "../../../../../domain/LoginService";
 import { Controller, HttpResponse } from "../../../../../domain/Controller";
