+++ src/presentation/controllers/users/CreateUserController.ts
@@ -1,5 +1,5 @@
 import { Request } from "express";
-import { HttpStatusCode } from "../../../infraestructure/utils/HttpStatusCode";
+import { HttpStatusCode } from "../../../domain/HttpStatusCode";
 import { z } from "zod";
 import { ICreateUserService } from "../../../domain/CreateUserService";
 import { Controller, HttpResponse } from "../../../domain/Controller";
